Add climbing bias when exiting climbing to active state
If the rays could reach the bottom of the area you just climbed up from (e.g. big climbable stair) it just looked like a steep slope.

Now when entering the active state from the climbing state it will impulse the interest bias in a direction for a short time to help you get up the stair.